Refactor CsvConverterBuilder.kt -> CsvConverterCreator.kt and refactor into class rather than functions, the same way EntityFilesCreator is done, but without inner class. The file creation of CsvConverter has been moved to CsvConverterCreator rather than in RoomProcessor. Move functions in CsvExportFunSpecs.kt and CsvImportFunSpecs.kt into CsvConverterCreator.kt, both files will be removed.
